move alice and bob to protocol package
