+++ protocol/bob/instance.go
