+++ protocol/bob/net.go
