+++ protocol/bob/offers.go
