+++ protocol/bob/recovery.go
