+++ protocol/bob/recovery_test.go
