+++ protocol/bob/swap_state.go
